refactor: move FollowerCollection and SingleFollower components
