+++ web/components/ui/followers/FollowerCollection/FollowerCollection.module.scss
@@ -1,0 +1,3 @@
+.followers {
+	width: 100%;
+}
