+++ web/components/ui/followers/SingleFollower/SingleFollower.module.scss
@@ -31,7 +31,3 @@
     height: 100%;
   }
 }
-
-.followers {
-  width: 100%;
-}
